Remove conditional which can never be reached.
The Form::Functional->process method always always expects an array back
as it'll go on to dereference it, which would cause everything to explode.

Also fix calling ->process with no args, which ends up implicitly passing
(undef) as the values (ergo the condition never being able to be reached)..

I'm not particularly happy with this - should we instead explode with an error
message when no values at all are passed in?